Initial Commit of "Find the Rio IP"
This program will only work on a Raspberry Pi or Linux based system.
This program will find the IP of the RoboRio, and save it into a local text file. It starts whenever the Pi is turned on and executes Linux shell commands to find all the devices on the subnet, and finds which device's IP is the Rio's by comparing the Rio's MAC address to all the MAC Addresses and IP's that are returned from the shell commands. I plan to change the vision code to get the IP to use from the text file that is created.